Study-01/requirements.txt
tensorflow==2.20.0
numpy==2.4.2
pillow==12.1.1
